+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: mvan-der <mvan-der@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/08/31 10:29:05 by mvan-der      #+#    #+#                 */
-/*   Updated: 2020/11/21 09:51:37 by mvan-der      ########   odam.nl         */
+/*   Updated: 2020/11/26 10:07:32 by mvan-der      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	unsigned char	*a;
 	unsigned char	*b;
-	unsigned int	i;
+	size_t			i;
 
 	a = (unsigned char*)s1;
 	b = (unsigned char*)s2;
